Add files via upload
In the JS section of this Codepen is a React class component called App that is rendered to the view. For the following challenges, you will only need to edit code in the JS section, not the HTML or CSS.

Challenge 1: Add a fourth country to the list stored in App's state.
Challenge 2: Cause the continent to show up next to each country name in the view, as in: "Laos - Asia"
Challenge 3: There is a CSS class called "title" defined in the CSS section. Apply this class to the h1 element for "My Travel Bucket List".
NOTE: When React talks about "lists and keys", it does not necessarily mean an HTML listitem element (<li>). It could be a list composed from <span> elements, <div> elements, et cetera, typically rendered from multiple items in an array which are all treated to the same markup. In this Codepen example, <div> is being used.